Make packages generic
The JSON.Types package has been made generic so that the precise integer
and float types can be defined. The package can be instantiated as
follows:

   package Types is new JSON.Types (Long_Integer, Long_Float);

Optionally, one can add:

   use Types;

The JSON.Parsers and JSON.Tokenizers packages require an instantiation
of the JSON.Types package as their only generic parameter. For example,
the JSON.Parsers package can be instantiated as follows:

   package Parsers is new JSON.Parsers (Types);

where Types is the instantiated package above.

Signed-off-by: onox <denkpadje@gmail.com>