Modify how parsed JSON values are stored to speed up parsing
A parser will now use a memory allocator to store JSON values of a
JSON array or the key-value pairs of a JSON object in an array of
vectors. The number of vectors does not depend on the size of the parsed
JSON text, but depends on the maximum nesting depth, specified when
instantiating JSON.Types.

JSON_Value no longer contains an Unbounded_String but instead contains
the offset and length of the string in the original stream. This avoids
copying the string and allocating many Unbounded_Strings.

A speed up of more than 4x compared to v2.0.0 or 17x compared to v1.0.0
has been achieved.

Subprogram

 - function Get
     (Object  : JSON_Value;
      Key     : String;
      Default : String) return JSON_Value

has been removed because a JSON string value refers to an offset in the
original stream.

Subprograms

 - function Contains (Object : JSON_Value; Key : String) return Boolean

 - function Get (Object : JSON_Value; Key : String) return JSON_Value

now have a time complexity of O(n) instead of O(1).

Package JSON.Parsers now has a generic parameter Check_Duplicate_Keys
which can be used to raise a Constraint_Error during parsing if an
object contains duplicate keys.

Signed-off-by: onox <denkpadje@gmail.com>